+++ proxy-router/mobile/storage.go
@@ -1,8 +1,14 @@
 package mobile
 
 import (
+	"errors"
 	"fmt"
 	"sync"
+)
+
+var (
+	ErrKeyNotFound = errors.New("key not found")
+	ErrKeyExists   = errors.New("key already exists")
 )
 
 // MemoryKeyValueStorage implements interfaces.KeyValueStorage in-memory.
@@ -23,7 +29,7 @@ func (m *MemoryKeyValueStorage) Get(key string) (string, error) {
 	defer m.mu.RUnlock()
 	v, ok := m.data[key]
 	if !ok {
-		return "", fmt.Errorf("key not found: %s", key)
+		return "", fmt.Errorf("%w: %s", ErrKeyNotFound, key)
 	}
 	return v, nil
 }
@@ -32,7 +38,7 @@ func (m *MemoryKeyValueStorage) Insert(key string, value string) error {
 	m.mu.Lock()
 	defer m.mu.Unlock()
 	if _, exists := m.data[key]; exists {
-		return fmt.Errorf("key already exists: %s", key)
+		return fmt.Errorf("%w: %s", ErrKeyExists, key)
 	}
 	m.data[key] = value
 	return nil
@@ -49,7 +55,7 @@ func (m *MemoryKeyValueStorage) Delete(key string) error {
 	m.mu.Lock()
 	defer m.mu.Unlock()
 	if _, exists := m.data[key]; !exists {
-		return fmt.Errorf("key not found: %s", key)
+		return fmt.Errorf("%w: %s", ErrKeyNotFound, key)
 	}
 	delete(m.data, key)
 	return nil
@@ -60,4 +66,13 @@ func (m *MemoryKeyValueStorage) DeleteIfExists(key string) error {
 	defer m.mu.Unlock()
 	delete(m.data, key)
 	return nil
+}
+
+// Clear removes all entries from the storage (best-effort secret wipe on shutdown).
+func (m *MemoryKeyValueStorage) Clear() {
+	m.mu.Lock()
+	defer m.mu.Unlock()
+	for k := range m.data {
+		delete(m.data, k)
+	}
 }
